Move bootstrap to right location.
